+++ packages/expo/src/ImagePicker/ExponentImagePicker.ts
@@ -1,0 +1,2 @@
+import { NativeModules } from 'react-native';
+export default NativeModules.ExponentImagePicker;
