+++ packages/expo/src/ImagePicker/ExponentImagePicker.web.ts
@@ -1,0 +1,5 @@
+export default {
+  get name(): string {
+    return 'ExponentImagePicker';
+  },
+};
